The R for MassSpectrometry meta-package

RforMassSpectrometry is a meta-package, to manage and document the RforMassSpectrometry project. The goal of the project to provide efficient, thoroughly documented, tested and flexible R software for the analysis and interpretation of high throughput mass spectrometry assays, including proteomics and metabolomics experiments. The project formalises the longtime collaborative development efforts of its members under the RforMassSpectrometry organisation to facilitate dissemination and accessibility of their work.

Details

See https://www.rformassspectrometry.org/ for details.

See also

Author

Maintainer: Laurent Gatto laurent.gatto@uclouvain.be (ORCID)

Authors:

Laurent Gatto laurent.gatto@uclouvain.be (ORCID)

Johannes Rainer Johannes.Rainer@eurac.edu (ORCID)

Sebastian Gibb mail@sebastiangibb.de (ORCID)